requirements.txt
Django==1.9.7
django-filter==0.13.0
gunicorn==19.6.0
mysqlclient==1.3.7
